% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beetdata.R
\docType{data}
\name{beet}
\alias{beet}
\title{Beet data for Example 2}
\format{An object of class \code{data.frame} with 15 rows and 3 columns.}
\usage{
data(beet)
}
\description{
Petersen (1994, p. 125) describes an experiment conducted to assess the effects
of five different quantities of N-fertiliser (0, 35, 70, 105 and 140 kg N/ha) on root dry
matter yield of sugar beets (t/ha) with three complete replications laid out in three
randomized complete blocks. One objective of this experiment is to determine the amount
of fertilizer maximizing yield.
Petersen, R.G. (1994). Agricultural field experiments. Design and analysis. New York:
Marcel Dekker.
}
\keyword{data}
