% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example3.R
\docType{package}
\name{example3}
\alias{example3}
\alias{example3-package}
\title{EXAMPLE 3 Polynomial regression model for an experiment with two quantitative level treatment factors}
\description{
(Gomez & Gomez, 1984, p. 401): Nitrogen uptake (g/pot) of rice was studied in a two-factor
greenhouse experiment involving duration of water stress (W) and level of nitrogen application (N).
The experiment had four water-stress levels (0, 10, 20 and 40 days) as main-plot treatments and
four nitrogen rates (0, 90, 180 and 270 kg/ha) as sub-plot treatments. The main plots were
randomized in four complete blocks. Treatment factors in combination often interact and polynomial
models for two or more quantitative treatment factors will usually require the inclusion of polynomial
interaction effects. Polynomial models are built by adding individual treatment
effects in a strict order of importance where the lower-degree effects of a
factorial combination must be added first before the higher-degree effects
 of that factorial combination.
}
\details{
Code

The example code shown below can be copied and pasted into any R console terminal window and
will provide a basic analysis of the greenhouse rice experiment data (type greenrice to view the data).

Graphical output should appear in the gui graphics window but, if required, it can be diverted to
a suitable pdf file by using a pdf file command. Opening pdf() and closing dev.off() commands are
shown hashed-out in the example code but can be copied and pasted without the hashes, if required.

Similarly, textual output should appear in the gui terminal window but, if required, it can be diverted to
a suitable text file by using a sink file command. Opening and closing sink() commands are
shown hashed-out in the example code but can be copied and pasted without the hashes, if required.

 Analysis
More to do!!
}
\examples{
## Copy and paste the following code into a R console or GUI to run examples
## Packages lmerTest, lattice and pbkrtest MUST be installed

\dontrun{
require(lmerTest)
require(lattice)
require(pbkrtest)

data(greenrice)
greenrice$loguptake=log(greenrice$uptake)
PolW=poly(greenrice$W, degree=2, raw=TRUE)
colnames(PolW)=c("linW","quadW")
PolN=poly(greenrice$N, degree=2, raw=TRUE)
colnames(PolN)=c("linN","quadN")
greenrice=cbind(greenrice,PolW,PolN)

## untransformed uptake data stratified anova
greenrice.uptake=aov(uptake~Replicate+linW*linN+quadN+quadW+Error(Replicate/Main),greenrice)
summary(greenrice.uptake, ddf="Kenward-Roger",type = 1)

## log transformed uptake data stratified anova
greenrice.loguptake=aov(loguptake~Replicate+linW*linN+quadN+quadW+Error(Replicate/Main),greenrice)
summary(greenrice.loguptake, ddf="Kenward-Roger",type = 1)

greenrice.uptake = lmer(uptake ~ Replicate + N*W  + (1|Replicate:Main) , data=greenrice)
greenrice.loguptake= lmer(loguptake ~ Replicate + N*W  + (1|Replicate:Main), data=greenrice)

## fitted quadratic loguptake curve versus nitrogen rate treatments
panel.plot <- function(x, y) {
panel.xyplot(x, y) # show points
linW=c(0,10,20,40)[panel.number()]
panel.curve(-1.16+0.0176*linW-0.00116*linW*linW +0.0068*x-0.00000938*x*x-0.0000907*linW*x,
from=0,to=270, type="l", lwd=2)
}
xyplot(loguptake ~ linN|factor(linW),data=greenrice,
 main="Fig 4: Marginal model for nitrogen rate",
 xlab = " Nitrogen (kg/ha)", ylab = "Log nitrogen uptake (g/pot)",
 strip = strip.custom(strip.names = TRUE, strip.levels = TRUE),
panel = panel.plot)

## fitted quadratic loguptake curve versus water stress treatments
panel.plot <- function(x, y) {
panel.xyplot(x, y) # show points
linN=c(0,90,180,270)[panel.number()]
panel.curve(-1.16+0.0176*x-0.00116*x*x +0.0068*linN-0.00000938*linN*linN-0.0000907*x*linN,
from=0,to=40, type="l", lwd=2)
}
xyplot(loguptake ~ linW|factor(linN),data=greenrice,
 main="Fig 4: Marginal model for water stress",
 xlab = " Water stress (days)", ylab = "Log nitrogen uptake (g/pot)",
 strip = strip.custom(strip.names = TRUE, strip.levels = TRUE),
panel = panel.plot)

## shows residuals from untransformed versus log transformed uptake data
par(mfrow=c(1,2),oma=c(0,0,2,0))
a=plot(greenrice.uptake,main="Example 3: untransformed", ylab="Residuals N uptake ")
b=plot(greenrice.loguptake,main="Example 3: transformed ", ylab="Residuals log N uptake")
print(a, position = c(0, 0, 0.5, 1), more = TRUE)
print(b, position = c(0.5, 0, 1, 1))

## lme4 analysis of log uptake assuming a quadratic response surface model for water and nitrogen
greenrice.lmer0=lmer(loguptake~linN*linW+quadN+quadW+(1|Replicate)+
(1|Replicate:Main),data=greenrice)
anova(greenrice.lmer0, ddf="Kenward-Roger",type = 1)
summary(greenrice.lmer0, ddf="Kenward-Roger",type = 1)

}

}
\references{
Gomez, K.A., & Gomez, A.A. (1984). Statistical procedures for agricultural research, 2nd edn. New York: Wiley.
}
