% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenricedata.R
\docType{data}
\name{greenrice}
\alias{greenrice}
\title{Rice data for Example 3}
\format{An object of class \code{data.frame} with 64 rows and 5 columns.}
\usage{
data(greenrice)
}
\description{
Gomez & Gomez (1984, p. 401): Nitrogen uptake (g/pot) of rice was studied in a two-factor greenhouse experiment involving duration of water stress (W) and level of nitrogen application (N). The experiment had four water-stress levels (0, 10, 20 and 40 days) as main-plot treatments and four nitrogen rates (0, 90, 180 and 270 kg/ha) as sub-plot treatments. The main plots were randomized in four complete blocks.
}
\details{
\itemize{
       \item{uptake} Nitrogen uptake (g/pot)
       \item{W} Duration of water stress
       \item{N} Level of nitrogen application 
       \item{Replicate} Replicate
       \item{Main} Main
       }
}
\examples{
data(greenrice)

}
\references{
Gomez, K. A., & Gomez, A. A. (1984). \emph{Statistical Procedures for Agricultural Research}. 
 John Wiley & Sons.
}
\keyword{datasets}
