% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricedata.R
\docType{data}
\name{rice}
\alias{rice}
\title{Rice data for Example 1}
\format{An object of class \code{data.frame} with 135 rows and 8 columns.}
\usage{
data(rice)
}
\description{
Gomez & Gomez (1984, p. 143) report a rice experiment with three management practices
(minimum, optimum, intensive), five different amounts of nitrogen (N) fertilizer
(0, 50, 80, 110, 140 kg/ha), and three varieties (V1, V2, V3). The experiment
involved variety and management as qualitative treatment factors and nitrogen
fertilizer as a quantitative treatment factor. Overall, there were 45 treatment
#'Gomez, K.A., & Gomez, A.A. (1984). Statistical procedures for agricultural research,
2nd edn. New York: Wiley.
}
\keyword{data}
