% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriTutorial-package.r
\docType{package}
\name{agriTutorial}
\alias{agriTutorial}
\alias{tutorial}
\alias{agriTutorial-package}
\title{Tutorial analysis of designed agricultural experiments}
\description{
The \code{agri.tutorial} package is intended as a teaching aid and
supports a tutorial approach to the analysis of experimental data for research workers.
The package examines a number of real data sets from practical experiments and explores
the analysis of these data sets using modern statistical methodolgy.
}
\details{
i) \code{\link[agriTutorial]{example1}} : The \code{example1} page provides a tutorial analysis of a split-plot design
with one quantitative treatment factor and one qualitative treatment factor.

ii) \code{\link[agriTutorial]{example2}} : The \code{example2} page provides a tutorial analysis of block design
with one qualitative treatment factor.

iii) \code{\link[agriTutorial]{example3}} : The \code{example3} page provides a tutorial analysis of a response surface
model for an experiment with two quantitative treatment factors.

iv) \code{\link[agriTutorial]{example4}} : The \code{example4} page provides a tutorial analysis of a single quantitative factor
with repeated measures in time.

v) \code{\link[agriTutorial]{example5}} : The \code{example5} page provides a tutorial analysis of transformation of
quantitative treatment levels.
}
\references{
H.P. Piepho1, R.N. Edmondson. (2017). A tutorial on the statistical analysis of factorial experiments with qualitative and quantitative
treatment factor levels.Journal of Agronomy and Crop Science. Submitted
}
