% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorghumdata.R
\docType{data}
\name{sorghum}
\alias{sorghum}
\title{Sorghum data for Example 4}
\format{An object of class \code{data.frame} with 100 rows and 7 columns.}
\usage{
data(sorghum)
}
\description{
Milliken & Johnson (1992, p. 429) describe an experiment with four sorghum varieties, in which the leaf area
index was assessed in five consecutive weeks starting two weeks after emergence. The experiment was laid out
in five randomized complete blocks. The observed data are plotted against week in Figure 5.
Milliken, G.A., & Johnson, D.E. (1992). Analysis of messy data. Volume I: Designed experiments. Boca Raton: CRC Press.
}
\keyword{data}
